Red Black Tree

Introduction

Red Black Tree is a self-balancing binary search tree. It is a binary search tree with one extra bit of information per node: the color, which is either red or black. By constraining the node colors on any simple path from the root to a leaf, red-black tree ensures that no such path is more than twice as long as any other, so that the tree is approximately balanced. A red-black tree is a binary search tree that satisfies the following red-black properties:

Every node is either red or black.

The root is black.

A red node cannot have a red child.

Every path from the root to a leaf contains the same number of black nodes.

API

SearchingTrees.RBTrees.RBTreeMap — Type

RBTreeMap{K, V}

A red-black tree map with key type K and value type V.

Fields

root::Union{RBTreeMapNode{K, V}, Nil}: the root node of the tree, might either be a RBTreeMapNode or nil when the tree is empty.

Constructors

RBTreeMap{K, V}(): create an empty tree map

An RBTreeMap is a subtype of AbstractDict{K, V} and it is an ordered map based on red-black tree data structure. It implements the following interfaces:

Base.get(d::RBTreeMap{K, V}, key::K, default) where{K, V}:

Base.getindex(d::RBTreeMap{K, V}, key::K) where{K, V}

Base.setindex!(d::RBTreeMap{K, V}, value::V, key::K) where{K, V}

Base.delete!(d::RBTreeMap{K, V}, key::K) where{K, V}

Base.iterate(d::RBTreeMap{K, V}) where{K, V}

Base.empty(d::RBTreeMap{K, V}) where{K, V}: return an empty RBTreeMap with the same key and value types as d

Base.empty!(d::RBTreeMap{K, V}) where{K, V}: empty d in place

Base.isempty(d::RBTreeMap{K, V}) where{K, V}: return true if d is empty

Base.get — Method

Base.get(d::RBTreeMap{K, V}, key::K, default)

Get the value of key in d, return default if key is not in d

Base.getindex — Method

Base.getindex(d::RBTreeMap{K, V}, key::K)

Get the value of key in d, throw KeyError if key is not in d. Called by d[key].

Base.setindex! — Method

Base.setindex!(d::RBTreeMap{K, V}, value::V, key::K)

Set the value of key in d to value. Called by d[key] = value.

Base.delete! — Method

Base.delete!(d::RBTreeMap{K, V}, key::K)

Delete key in d. Called by delete!(d, key).

Base.iterate — Method

Base.iterate(d::RBTreeMap{K, V})

Iterate through d in ascending order of keys.

Base.empty — Method

Base.empty(d::RBTreeMap{K, V})

Return an empty RBTreeMap with the same key and value types as d.

Base.empty! — Method

Base.empty!(d::RBTreeMap{K, V})

Empty d in place.